requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx>=0.25.0
requests>=2.31.0

# AI服务提供商（按需安装）
# OpenAI (如果需要)
openai>=1.12.0

# Google Gemini (如果需要)
# google-generativeai>=0.3.0

# Anthropic Claude (如果需要)
# anthropic>=0.18.0

